Refactor backend test specs for lint compliance
