+++ backend/src/tests/temp/database.spec.ts
@@ -2,17 +2,15 @@ import { describe, it, expect, beforeAll, afterAll } from 'vitest'
 import DatabaseConnection from '../database/connection.js'
 import { MigrationRunner } from '../database/migrations.js'
 
-describe('Database Connection and Migrations', () => {
-  beforeAll(async () => {
-    // Setup test database
-    process.env.DB_PATH = ':memory:' // Use in-memory database for testing
-  })
+beforeAll(() => {
+  process.env.DB_PATH = ':memory:'
+})
 
-  afterAll(() => {
-    // Cleanup
-    DatabaseConnection.close()
-  })
+afterAll(() => {
+  DatabaseConnection.close()
+})
 
+describe('Database Connection health', () => {
   it('should create database connection', () => {
     const db = DatabaseConnection.getInstance()
     expect(db).toBeDefined()
@@ -22,25 +20,32 @@ describe('Database Connection and Migrations', () => {
     const isHealthy = DatabaseConnection.isHealthy()
     expect(isHealthy).toBe(true)
   })
+})
 
+describe('Database migrations execution', () => {
   it('should run all migrations successfully', async () => {
     const migrationRunner = new MigrationRunner()
-    
-    // Should not throw an error
     await expect(migrationRunner.runMigrations()).resolves.toBeUndefined()
+  })
+})
+
+describe('Database schema verification', () => {
+  beforeAll(async () => {
+    const migrationRunner = new MigrationRunner()
+    await migrationRunner.runMigrations()
   })
 
   it('should have all required tables after migrations', () => {
     const db = DatabaseConnection.getInstance()
-    
+
     const tables = db.prepare(`
-      SELECT name FROM sqlite_master 
+      SELECT name FROM sqlite_master
       WHERE type='table' AND name NOT LIKE 'sqlite_%'
       ORDER BY name
     `).all() as { name: string }[]
 
     const tableNames = tables.map(t => t.name)
-    
+
     expect(tableNames).toContain('instruments')
     expect(tableNames).toContain('portfolio_positions')
     expect(tableNames).toContain('trades')
@@ -52,14 +57,14 @@ describe('Database Connection and Migrations', () => {
 
   it('should have proper indexes on instruments table', () => {
     const db = DatabaseConnection.getInstance()
-    
+
     const indexes = db.prepare(`
-      SELECT name FROM sqlite_master 
+      SELECT name FROM sqlite_master
       WHERE type='index' AND tbl_name='instruments'
     `).all() as { name: string }[]
 
     const indexNames = indexes.map(i => i.name)
-    
+
     expect(indexNames).toContain('idx_instruments_symbol')
     expect(indexNames).toContain('idx_instruments_esg')
     expect(indexNames).toContain('idx_instruments_vegan')
